api/requirements.txt
argcomplete==2.0.0
asgiref==3.7.2
Beaker==1.12.1
beautifulsoup4==4.12.2
blivet==3.7.1
Brlapi==0.8.4
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
distro==1.8.0
Django==4.2.4
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
gpg==1.17.1
humanize==3.13.1
idna==3.4
langtable==0.0.62
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
MarkupSafe==2.1.2
nftables==0.1
oauthlib==3.2.2
olefile==0.46
packaging==23.0
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.4.0
ply==3.11
productmd==1.36
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.23.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGObject==3.44.1
PyJWT==2.8.0
pykickstart==3.47
pyOpenSSL==21.0.0
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python3-openid==3.2.0
pytz==2023.3
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
regex==2023.6.3
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rpm==4.18.1
selinux==3.5
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sos==4.5.1
soupsieve==2.4.1
sqlparse==0.4.4
systemd-python==235
Tempita==0.5.2
urllib3==1.26.16
